Extracted tflint as separated test
